Split up editor export code into multiple files
