+++ editor/editor_run_native.cpp
@@ -30,9 +30,9 @@
 
 #include "editor_run_native.h"
 
-#include "editor/editor_export.h"
 #include "editor/editor_node.h"
 #include "editor/editor_scale.h"
+#include "editor/export/editor_export_platform.h"
 
 void EditorRunNative::_notification(int p_what) {
 	switch (p_what) {
