+++ editor/export/editor_export_platform_pc.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  export_plugin.h                                                      */
+/*  editor_export_platform_pc.h                                          */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,25 +28,55 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef LINUXBSD_EXPORT_PLUGIN_H
-#define LINUXBSD_EXPORT_PLUGIN_H
+#ifndef EDITOR_EXPORT_PLATFORM_PC_H
+#define EDITOR_EXPORT_PLATFORM_PC_H
 
-#include "core/io/file_access.h"
-#include "editor/editor_settings.h"
-#include "editor/export/editor_export_platform_pc.h"
-#include "platform/linuxbsd/logo.gen.h"
-#include "scene/resources/texture.h"
+#include "editor_export_platform.h"
 
-class EditorExportPlatformLinuxBSD : public EditorExportPlatformPC {
-	HashMap<String, String> extensions;
-	Error _export_debug_script(const Ref<EditorExportPreset> &p_preset, const String &p_app_name, const String &p_pkg_name, const String &p_path);
+class EditorExportPlatformPC : public EditorExportPlatform {
+	GDCLASS(EditorExportPlatformPC, EditorExportPlatform);
+
+private:
+	Ref<ImageTexture> logo;
+	String name;
+	String os_name;
+
+	int chmod_flags = -1;
 
 public:
-	void set_extension(const String &p_extension, const String &p_feature_key = "default");
-	virtual List<String> get_binary_extensions(const Ref<EditorExportPreset> &p_preset) const override;
+	virtual void get_preset_features(const Ref<EditorExportPreset> &p_preset, List<String> *r_features) override;
+
+	virtual void get_export_options(List<ExportOption> *r_options) override;
+
+	virtual String get_name() const override;
+	virtual String get_os_name() const override;
+	virtual Ref<Texture2D> get_logo() const override;
+
+	virtual bool can_export(const Ref<EditorExportPreset> &p_preset, String &r_error, bool &r_missing_templates) const override;
 	virtual Error export_project(const Ref<EditorExportPreset> &p_preset, bool p_debug, const String &p_path, int p_flags = 0) override;
-	virtual String get_template_file_name(const String &p_target, const String &p_arch) const override;
-	virtual Error fixup_embedded_pck(const String &p_path, int64_t p_embedded_start, int64_t p_embedded_size) override;
+	virtual Error sign_shared_object(const Ref<EditorExportPreset> &p_preset, bool p_debug, const String &p_path);
+	virtual String get_template_file_name(const String &p_target, const String &p_arch) const = 0;
+
+	virtual Error prepare_template(const Ref<EditorExportPreset> &p_preset, bool p_debug, const String &p_path, int p_flags);
+	virtual Error modify_template(const Ref<EditorExportPreset> &p_preset, bool p_debug, const String &p_path, int p_flags) { return OK; };
+	virtual Error export_project_data(const Ref<EditorExportPreset> &p_preset, bool p_debug, const String &p_path, int p_flags);
+
+	void set_extension(const String &p_extension, const String &p_feature_key = "default");
+	void set_name(const String &p_name);
+	void set_os_name(const String &p_name);
+
+	void set_logo(const Ref<Texture2D> &p_logo);
+
+	void add_platform_feature(const String &p_feature);
+	virtual void get_platform_features(List<String> *r_features) override;
+	virtual void resolve_platform_feature_priorities(const Ref<EditorExportPreset> &p_preset, HashSet<String> &p_features) override;
+
+	int get_chmod_flags() const;
+	void set_chmod_flags(int p_flags);
+
+	virtual Error fixup_embedded_pck(const String &p_path, int64_t p_embedded_start, int64_t p_embedded_size) {
+		return Error::OK;
+	}
 };
 
-#endif // LINUXBSD_EXPORT_PLUGIN_H
+#endif // EDITOR_EXPORT_PLATFORM_PC_H
