+++ platform/android/export/export.cpp
@@ -30,10 +30,10 @@
 
 #include "export.h"
 
-#include "export_plugin.h"
-
 #include "core/os/os.h"
 #include "editor/editor_settings.h"
+#include "editor/export/editor_export.h"
+#include "export_plugin.h"
 
 void register_android_exporter() {
 	EDITOR_DEF("export/android/android_sdk_path", "");
