+++ platform/ios/export/export_plugin.h
@@ -40,8 +40,8 @@
 #include "core/os/os.h"
 #include "core/templates/safe_refcount.h"
 #include "core/version.h"
-#include "editor/editor_export.h"
 #include "editor/editor_settings.h"
+#include "editor/export/editor_export_platform.h"
 #include "main/splash.gen.h"
 #include "platform/ios/logo.gen.h"
 #include "string.h"
