+++ platform/linuxbsd/export/export.cpp
@@ -30,6 +30,7 @@
 
 #include "export.h"
 
+#include "editor/export/editor_export.h"
 #include "export_plugin.h"
 
 void register_linuxbsd_exporter() {
